feat: expect, assert, and precondition functions
Adds functions to express expectations, assertions, and preconditions in
tests. Expectations are conditions that will cause a test to fail but to
continue running if the condition is not met. Assertions are conditions
that will cause a test to fail immediately if the condition is not met.
Preconditions are just assertions with different identifiers and failure
messages to easily distinguish whether a particular condition and
potential failure is part of the system under test, or precondition for
testing.