feat: add neq operations
Adds expect, assert, and precondition functions for expecting that
values are not equal to each other. Also fixes some existing tests that
had copy/pasted names which were incorrect, and improves some of the
expressions and test data so the suites center around a snigle
expectation which is true in the the success cases and false in the fail
case. Previously some of the tests expected a different condition in the
failure case than they did in the expect case; so even though the tests
were correct they required more effort to read.